feat: add microscopic value-level ARM as Step 4 of the pipeline
- Add microscopic_experiment_association_rules_values.py:
  value-level FP-Growth ARM anchored to the label-level rules
  produced by Step 3. Active labels are collected from all
  rules.csv files (rglob across experiment-label dirs); only
  LABEL=value items whose prefix is in the active set are kept.
  Transactions are aggregated per Sample_ID (union of CF-neighbour
  values) before encoding. Pipeline mirrors Step 3 exactly:
  auto-calibrated or manual support/confidence/lift grids,
  two-level loky parallelism (outer=k, inner=support), heatmaps,
  k_comparison cross-k summary.

- Optimise calibrate_parameters: replace O(n) linear scan with
  O(log n) binary search; use max_len=2 for calibration probes
  and max_len=4 for production FP-Growth calls.

- Fix numpy.bool_ crash: add .astype(float) to all item_supports
  derivation points (encoded_df.mean() returns bool dtype on some
  numpy versions, causing .round() / .to_dict() to fail).

- Replace pd.read_csv(rules_csv).empty with st_size < 120 in
  cleanup_empty_folders for O(1) empty-file detection.

- Extend main.py to support Step 4:
  - Register _SRC[4] and _STEP_NAMES[4].
  - Add _outputs_step4_exist() and _inputs_step4_exist() (checks
    both transactions_values.csv and at least one Step 3 rules.csv
    per region; distinct error messages for each missing input).
  - Add run_step4() with identical parameter surface to run_step3().
  - Update --steps choices to [1,2,3,4], default to all four.
  - Rename argparse group to "Steps 3 & 4: association rules".
  - Fix missing def for _outputs_step3_exist (orphan docstring bug).

+++ src/macroscopic_experiment_association_rules.py
@@ -956,7 +956,7 @@ def calibrate_parameters(
     """
     print('  > Calibrating parameters from item frequencies...')
 
-    item_supports = encoded_df.mean().sort_values()
+    item_supports = encoded_df.mean().astype(float).sort_values()
     print('  > Item supports:')
     for item, sup in item_supports.items():
         print(f'    {item}: {sup:.4f}')
@@ -1114,7 +1114,7 @@ def calibrate_parameters(
         'raw_lift_max':      round(raw_lift_max, 4),
         'min_conf_observed': round(min_conf_observed, 4) if min_conf_observed is not None else None,
         'max_conf_observed': round(max_conf_observed, 4) if max_conf_observed is not None else None,
-        'item_supports':     item_supports.round(4).to_dict(),
+        'item_supports':     item_supports.astype(float).round(4).to_dict(),
     }
 
     print(
@@ -1263,7 +1263,7 @@ def _process_one_k(
         return k, None, row
 
     df_encoded    = extract_labels(labels_csv)
-    item_supports = df_encoded.mean().sort_values()
+    item_supports = df_encoded.mean().astype(float).sort_values()
 
     if auto_calibrate:
         params = calibrate_parameters(
